# Fix #1818: Improve performance of Hashers.hashAsVersion