Before the change the performance of `testSuite/test:fastOpt` where:
```
[debug] Linker: Read info: 19896 us
[debug] Linker: Check Infos: 374083 us
[debug] Linker: Compute reachability: 193762 us
[debug] Linker: Assemble LinkedClasses: 1023952 us 
[debug] Linker: Check IR: 2641108 us
[debug] Linker: cache stats: reused: 5164 -- invalidated: 0 -- trees read: 0
```
After the change it has a 13X improvement:
```
[debug] Linker: Read info: 26999 us
[debug] Linker: Check Infos: 418228 us
[debug] Linker: Compute reachability: 1427003 us
[debug] Linker: Assemble LinkedClasses: 78869 us
[debug] Linker: Check IR: 2693044 us
[debug] Linker: cache stats: reused: 5164 -- invalidated: 0 -- trees read: 0
```