@sjrd 2015-07-31
Waw! That's amazing! 

Maybe use an indexed array for `hexDigit`?