@nicolasstucki 2015-08-10
Of all the options that where proposed, `Character.forDigit(digit, 16)` was the fastest wit 25X over the original and the most compact.